% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACM.R
\name{BACM}
\alias{BACM}
\title{A banding parameter estimator for banded autocovariance matrix}
\usage{
BACM(data, K, s)
}
\arguments{
\item{data}{observation vector}

\item{K}{a parameter that generate K by K  submatrix}

\item{s}{block size parameter}
}
\value{
estimated band
}
\description{
A banding parameter estimator for banded autocovariance matrix
}
