% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACM.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of R function (\code{BACM_R}) and Cpp functions (\code{BACM_C}).
}
