% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateARMA}
\alias{generateARMA}
\title{A ARMA series sampler using Rcpp}
\usage{
generateARMA(p, q, n, arParams, maParams)
}
\arguments{
\item{p}{order of autoregression}

\item{q}{order of moving average}

\item{n}{length of time series}

\item{arParams}{a vector of autoregression parameters}

\item{maParams}{a vector of moving average parameters}
}
\value{
the desired arma series
}
\description{
A ARMA series sampler using Rcpp
}
\examples{
\dontrun{
error<- generateARMA(3,4,1000,c(.5,-.2),.3)
}
}
