% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler}
\usage{
gibbsC(N, n, a, b)
}
\arguments{
\item{N}{number of samples}

\item{n}{number of trials}

\item{a}{non-negative parameters of the Beta distribution.}

\item{b}{non-negative parameters of the Beta distribution.}
}
\description{
A Gibbs sampler
}
